requirements.txt
Django
psycopg2
mysqlclient
python-dotenv
requests
pytest-django